+++ new_resources/command_books/hero.py
@@ -12,6 +12,7 @@ IMAGE_DIR = config.RESOURCES_DIR + '/command_books/hero/'
 class Key:
     # Movement
     JUMP = 'alt'
+    WORLD_MAP = 'pageup'
     FLASH_JUMP = 'alt'
     UP_JUMP = 'shift'
     ROPE = 'c'